# api-gateway/target/classes/openapi.yaml
openapi: 3.1.0
info:
  contact: {}
  description: SpaZar Api Documentation - Microservices Gateway
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  termsOfService: urn:tos
  title: SpaZar Online Api Documentation (Microservices)
  version: "2.0"
servers:
- url: //localhost:8080/
...
paths:
...
  /api/v1/customer/login:
    post:
      summary: Customer login
      tags: [User authentication Api]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT